Changed buildTask init param 'actor' to UAC, also removing WH param
to start we dont need to wh param any more as we can tell by the uac access level
this means that we only need on build lambda funtion for building.
also updated the webhook :)

I have noticed that build is now broken, as it uses the commonProject to get and
update build and project info. But this no longer makes any sence. Also it was
there was a planed changed to lock the file throught the update process, which
commonProjects could not support anyway.